tools/gcp-filestore-monitor/gcp-filestore-resize/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
Flask==2.0.2
functions-framework==3.0.0
google-api-core==2.4.0
google-auth==2.6.0
google-cloud-access-context-manager==0.1.9
google-cloud-asset==3.7.1
google-cloud-common==1.0.0
google-cloud-filestore==1.0.0
google-cloud-org-policy==1.2.1
google-cloud-os-config==1.9.0
google-cloud-resource-manager==1.3.3
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
watchdog==1.0.2
Werkzeug==2.0.2
